refactor: move Medication entity to medication package
