+++ src/main/java/com/springboot/api/counselsession/entity/MedicationRecordHist.java
@@ -1,5 +1,6 @@
 package com.springboot.api.counselsession.entity;
 
+import com.springboot.api.medication.entity.Medication;
 import java.time.LocalDate;
 
 import org.hibernate.annotations.OnDelete;
