+++ src/main/java/com/springboot/api/medication/repository/MedicationRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.springboot.api.counselsession.entity.Medication;
+import com.springboot.api.medication.entity.Medication;
 
 public interface MedicationRepository extends JpaRepository<Medication, String> {
         List<Medication> findByItemNameContainingIgnoreCase(String keyword);
